backend/requirements.txt
fastapi==0.104.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
SQLAlchemy==2.0.23
uvicorn==0.24.0.post1

